requirements.txt
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
whitenoise==6.8.2
pillow==11.0.0